final push?
Been having a lot of issues with some of my changes so I'm pushing before the inevitable laptop crash that's been fighting me every time I work on this. My home page is simply refusing to cooperate, but I think the portfolio page turned out nicely and as I envisioned it!
I honestly couldn't think of much to alter the layout of my home page, but I did make a change or two at the large screen size. We'll see if they stick.